introduce treap based taversal
We use a treap as a kind of priority queue for scheduled node.

Compared to other solutions, states are minimized within the traversal
structure. We need only retain the schedule itself. No more
"current_rank_" flag var and the likes.

Furthermore, note that this structure makes no use of any bucket related
parameter. There is no need to keep track of the size of our dag.