detail:nodes: add schedule_or_send_down_method
Since only merge node are liable to have a parent count greater than
one, they're the only one we need to potentially schedule.

In this way, the send_down traversal can remain more or less the same
when the DAG doesn't have any merge node. It will effectively perform a
DFS traversal as before.

When merge nodes are present, the traversal will schedule the merge node
and stop the traversal until it can go back to that node, as determined
by the traversal object.