detail:traversal: make scheduling opportunistic
The previous scheduling method would schedule every node before visiting
them. This meant that there was always something in current_rank_ + 1,
provided that some node in the current rank had actually changed value.

With opportunistic scheduling, this is no longer the case. Consider the
DAG

A ---> B ---> C \
|                \
-------------------> D

and suppose B does not change as a result of A changing.

Then D, of rank 3, will be scheduled after having "send_down" from A,
even though the current_rank_ after A's visit will be 1. (The example is
admittedly not minimal as C could be removed.)

Thus, in the non intrusive traversal implementation, we need to keep
track of the "last_rank_" (to visit!) in order to ensure that the traversal
will remember to hit scheduled node.

However, in the intrusive case, we can leverage the efficiency of the
removal operations to simply look whether or not there are elements left
in the schedule. Note that this further allow the elimination of
"clear()" instruction at the end of the visit, as the schedule has to be
empty by the end.