benchmark: add a "god chain" benchmark and improve treap
This benchmark was designed to challenge the first implementation of the
treap.

This chain degraded the performance of treap by forcing it to perform to
many comparisons when scheduling merged node. Whereas other multimap
based strategies could simply schedule the node at its rank, the treap
had to find where to put the node in the schedule. Because of the way
the god chain is structured, this would always be at the end.

The situation was markedly improved by leveraging the fact that we can
expect that nodes of higher rank will be created and added to after
nodes of lower rank, and thus that the vector of children within a node
are _probably_ gonna be ordered, the highest rank being at the back.

We can use this observation to give a treap a hint as to where to place
a node to schedule.