benchmark: add random dag benchmarks
Try to generate a random dag parametrized in the following way:

1. N := node count, self explanatory
2. E := entropy, a measure of "chaos" in value propagation
3. M := merge_node_factor, the % of node in the dag being merged node

The random dag generation process adds node in one of two ways. It can
add a new, single parent node, by randomly selecting a previously
added node as a parent. Otherwise, it adds a merge node with two
parents, also by selecting these randomly from the set of previously
added node.

The nodes are all boolean and update their values, if visited during
traversal, by shaking a magic eight ball with the given entropy factor.
The higher the entropy, the less stable the value. Maximal entropy
produces a situation in which there is a 50% chance of getting "true".
Minimal entropy guarantees the node value is always set to "false".