define node schedule object
Given a directed acyclic graph with a single root, the rank number of a node is
defined as follow:

1. If the node is the root, then its rank number is 0
2. Otherwise, the rank number is the max rank number of the node's
   parents + 1

For a DAG having a single root node, the node schedules satisfy the
following:

1. There is exactly 1 node schedule per rank in the graph

Schedule object are organized as a singly linked list. When creating a
new node with no parent, it is assume to be root and a schedule of rank
0 is created. This node points to the schedule.

When adding a node with one or more parents, this new node will either
find its rank's schedule already created in the linked list and point to
it, or a new schedule corresponding to the node's rank will be added to
the list, and the node will point to it.

During graph traversal, node of a given rank can be added in the rank's
schedule as member of a list. Since nodes within the same rank need
not be sorted relative to one another, a list with O(1) insertion time is
optimal.

FIX schedule definition